Conflicts:
	laravel/form.php

+++ laravel/file.php
@@ -69,6 +69,30 @@ class File {
 	}
 
 	/**
+	 * Move a file to a new location.
+	 *
+	 * @param  string  $path
+	 * @param  string  $target
+	 * @return void
+	 */
+	public static function move($path, $target)
+	{
+		return rename($path, $target);
+	}
+
+	/**
+	 * Copy a file to a new location.
+	 *
+	 * @param  string  $path
+	 * @param  string  $target
+	 * @return void
+	 */
+	public static function copy($path, $target)
+	{
+		return copy($path, $target);
+	}
+
+	/**
 	 * Extract the file extension from a file path.
 	 * 
 	 * @param  string  $path
@@ -174,6 +198,18 @@ class File {
 	}
 
 	/**
+	 * Create a new directory.
+	 *
+	 * @param  string  $path
+	 * @param  int     $chmod
+	 * @return void
+	 */
+	public static function mkdir($path, $chmod = 0777)
+	{
+		return ( ! is_dir($path)) ? mkdir($path, $chmod, true) : true;
+	}
+
+	/**
 	 * Move a directory from one location to another.
 	 *
 	 * @param  string  $source
@@ -183,7 +219,7 @@ class File {
 	 */
 	public static function mvdir($source, $destination, $options = fIterator::SKIP_DOTS)
 	{
-		static::cpdir($source, $destination, true, $options);
+		return static::cpdir($source, $destination, true, $options);
 	}
 
 	/**
@@ -197,7 +233,7 @@ class File {
 	 */
 	public static function cpdir($source, $destination, $delete = false, $options = fIterator::SKIP_DOTS)
 	{
-		if ( ! is_dir($source)) return;
+		if ( ! is_dir($source)) return false;
 
 		// First we need to create the destination directory if it doesn't
 		// already exists. This directory hosts all of the assets we copy
@@ -221,7 +257,7 @@ class File {
 			{
 				$path = $item->getRealPath();
 
-				static::cpdir($path, $location, $delete, $options);
+				if (! static::cpdir($path, $location, $delete, $options)) return false;
 
 				if ($delete) @rmdir($item->getRealPath());
 			}
@@ -231,22 +267,25 @@ class File {
 			// files with the same name.
 			else
 			{
-				copy($item->getRealPath(), $location);
+				if(! copy($item->getRealPath(), $location)) return false;
 
 				if ($delete) @unlink($item->getRealPath());
 			}
 		}
 
 		if ($delete) rmdir($source);
+		
+		return true;
 	}
 
 	/**
 	 * Recursively delete a directory.
 	 *
 	 * @param  string  $directory
+	 * @param  bool    $preserve
 	 * @return void
 	 */
-	public static function rmdir($directory)
+	public static function rmdir($directory, $preserve = false)
 	{
 		if ( ! is_dir($directory)) return;
 
@@ -267,7 +306,18 @@ class File {
 			}
 		}
 
-		@rmdir($directory);
+		if ( ! $preserve) @rmdir($directory);
+	}
+
+	/**
+	 * Empty the specified directory of all files and folders.
+	 *
+	 * @param  string  $directory
+	 * @return void
+	 */
+	public static function cleandir($directory)
+	{
+		return static::rmdir($directory, true);
 	}
 
 	/**
